% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cluster_relabelling.R
\name{as_hcluslabels}
\alias{as_hcluslabels}
\alias{is_hcluslabels}
\title{as_hcluslabels}
\usage{
as_hcluslabels(x)

is_hcluslabels(x)
}
\arguments{
\item{x}{input should be convertible to character matrix such that
rows are labelled observations while columns contain hierarchical labels
of increasing complexity.}
}
\description{
Convert/test input to a standard format for cluster relabelling with
    \code{\link{relabel_hcluslabels}}.
}
\section{Functions}{
\itemize{
\item \code{is_hcluslabels()}: Check if object is in a valid hcluslabels format

}}
