% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cluster_relabelling.R
\name{check_relabelling}
\alias{check_relabelling}
\title{check_relabelling}
\usage{
check_relabelling(new, old = NULL)
}
\arguments{
\item{new}{a matrix of recoded labels, the result of relabel_hcluslabels(old)}

\item{old}{an optional matrix of raw labels}
}
\value{
unchanged labels
}
\description{
Given a hcluslabels matrix check whether the relabelling was successful.
}
