% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_levels}
\alias{get_levels}
\title{get_levels}
\usage{
get_levels(x, drop = FALSE, locale_order = FALSE)
}
\arguments{
\item{x}{input, a factor or character vector.}

\item{drop}{boolean - should levels not in the data be discarded?}

\item{locale_order}{boolean - should created factor levels be alphabetical?
Note the default option value (FALSE) is to order as in data
which differs from base R's as.factor, but agrees with
forcats::as_factor. For base-R as.factor equivalence use TRUE.}
}
\description{
Will extract factor levels - or, if input does not have levels,
    appropriate ones will be created.
}
\details{
Default R behaviour (as.factor) for a character vector would be to produce
alphabetical levels. This can vary by locale. Alternatively we can use
the data input order which will not vary by locale. This latter approach is
used by forcats::as_factor.

Logical vectors will always be assigned c(FALSE, TRUE).
}
