% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{glance_df}
\alias{glance_df}
\title{glance_df}
\usage{
glance_df(x, ...)
}
\arguments{
\item{x}{A data.frame}

\item{...}{Additional arguments for other methods.}
}
\value{
A single row summary "glance" of the data.frame.
    A tibble reporting nrow, ncol, complete rows and missing fraction.
}
\description{
A single row summary "glance" of a data.frame
}
\details{
This function has been "rescued" from \code{\link[broom]{broom}} where
    the glance tidier for data.frames is being deprecated.
}
\examples{
# example code
glance_df(mtcars)
}
