% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_misc.R
\name{lmelist_pooled_rsquaredglmm}
\alias{lmelist_pooled_rsquaredglmm}
\title{lmeList_pooled_r.squaredGLMM}
\usage{
lmelist_pooled_rsquaredglmm(mlist)
}
\arguments{
\item{mlist}{A list of linear mixed effects models
resulting from multiple imputation}
}
\value{
A matrix of r2 values from \code{\link[MuMIn]{r.squaredGLMM}}
}
\description{
This function calculates pooled r2 values for multiply-imputed
linear mixed effects models.
}
\details{
The \code{\link[MuMIn]{r.squaredGLMM}} function from MuMIn package
provides r2 values for GLMM such as those produced by
\code{\link[lme4]{lmer}}, \code{\link[lme4]{glmer}},
and \code{\link[nlme]{lme}}.
However, this is not a default method and r2 is not yet pooled
when analysing multiply-imputed mixed effects models with mice
i.e. \code{mice::pool.r.squared(my_mira)}
This function applies the custom r2 function to each model and uses the
Harel (2009) fisher-r-to-z method for pooling r2 values.
}
\references{
Harel, O. (2009).
The estimation of R2 and adjusted R2 in incomplete data sets using
multiple imputation. Journal of Applied Statistics, 36(10), 1109-1118.
}
