% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cluster_relabelling.R
\name{lsap_relabel_pair}
\alias{lsap_relabel_pair}
\title{lsap_relabel_pair}
\usage{
lsap_relabel_pair(new, old)
}
\arguments{
\item{new}{A vector of labels for subjects, to be modified such that labels
in new are matched to best fitting labels in old.}

\item{old}{A vector of labels for subjects, values of old will be used to
relabel new.}
}
\value{
A copy of \code{new} with the updated labels.
}
\description{
Function creates and applies a mapping from new -> old labels using
    a linear sum assignment solver (LSAP, see:
    \code{\link[clue]{solve_LSAP}}).
    For application to a set of hierarchical cluster labels see
    \code{\link{relabel_hcluslabels}}.
}
\examples{
lsap_relabel_pair(c("A", "B", "C"), c("x", "y", "z"))
}
