% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_stars}
\alias{make_stars}
\title{make_stars}
\usage{
make_stars(
  pvals,
  symbols = c("***", "**", "*"),
  cuts = c(0.001, 0.01, 0.05),
  nonsymbol = NA_character_,
  nablank = FALSE,
  check = TRUE
)
}
\arguments{
\item{pvals}{vector of p-values.}

\item{symbols}{symbols to use}

\item{cuts}{upper (right) bounds for each symbol (must be in ascending order)}

\item{nonsymbol}{return value for p-values not assigned a symbol in cuts.
default: NA.}

\item{nablank}{boolean - should NA values be returned as an empty string:
""(TRUE), or be passed as an NA (FALSE; default)?}

\item{check}{boolean - should pvals be checked for range?}
}
\description{
Produce "significance" stars for p-values.
    Defaults correspond to R's stats::summary.lm table but without the '.'
    for p-values 0.05 < p < 0.1.
}
\details{
Note: because the value of the last element of \code{cuts} is the largest
    value that will be assigned a symbol this can be set to 1 to assign
    a symbol to non-significant p-values.
}
\examples{
{
  make_stars(10^c(seq(from = -7, to = 0, by = 0.25)))
  # try a custom scheme where you get increasing stars the more zeros you have
  #      (with a single "*" for 0.01 < p <= 0.05)
  custom_cuts <- 10^c(seq(from = -10, to = -2, by = 1), log10(0.05))
  custom_symbols <- sapply(10:1, function(x) paste0(rep("*", x), collapse = ""))
  make_stars(10^c(seq(from = -7, to = 0, by = 0.25)),
             cuts = custom_cuts, symbols = custom_symbols)
}
}
\seealso{
stats::symnum
}
