% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{not_in}
\alias{not_in}
\alias{\%!in\%}
\title{not in}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\value{
A logical vector, indicating if each element of x was NOT found in
    table, values are TRUE or FALSE and never NA.
}
\description{
utility function, infix operator to negate \code{\link{\%in\%}}.
    This simplifies the common "\code{! x \%in\% y}" construction to
    "\code{x \%!in\% y}"
}
