% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cluster_relabelling.R
\name{relabel_cluster_pair}
\alias{relabel_cluster_pair}
\title{relabel_cluster_pair}
\usage{
relabel_cluster_pair(old, new)
}
\arguments{
\item{old}{A vector of labels for subjects, values of old will be used to
relabel new.}

\item{new}{A vector of labels for subjects to be matched to new.}
}
\value{
A copy of \code{new} with the matched labels from old.
}
\description{
Function creates a mapping from new -> old labels based on a greedy
    heuristic (see \code{\link{relabel_hcluslabels}} which implements for
    a hierarchical set of clusters.)
}
\examples{
relabel_cluster_pair(c("A", "B", "C"), c("x", "y", "z"))
}
