% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{selfname}
\alias{selfname}
\title{selfname}
\usage{
selfname(x, append = "none", prefix = "", suffix = "", sep = "")
}
\arguments{
\item{x}{an object (vector, list) with elements coercible to character (and
sensible to use as names).}

\item{append}{"none", "before", "after" options for appending contents of
x to existing names.}

\item{prefix}{add this before elements in names}

\item{suffix}{add this after elements in names}

\item{sep}{separator to use for constructed names}
}
\value{
a named object x
}
\description{
Use the elements of an object as their own names. This code extends
    \code{\link[stats]{setNames}}. Mostly useful for making iterable objects
    on the fly.
}
\examples{
{
    x <- 1:3
    y <- letters[1:3]
    xn <- stats::setNames(x,y)

    selfname(x)

    selfname(y)

    selfname(x, prefix = "v")

    selfname(x, prefix = y)

    selfname(xn, append = "after")
}


}
