% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_cluster_relabelling.R
\name{as_hcluslabels}
\alias{as_hcluslabels}
\alias{is_hcluslabels}
\title{as_hcluslabels}
\usage{
as_hcluslabels(x)

is_hcluslabels(x)
}
\arguments{
\item{x}{input should be convertable to character matrix}
}
\description{
Convert/test input standard format for cluster relabeling.
}
\section{Functions}{
\itemize{
\item \code{is_hcluslabels()}: Check if object is in a valid hcluslabels format

}}
